Refactor project structure, optimize file organization, and update imports
