+++ src/components/common/Header.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import './Header.css';
-import logo from '../assets/images/Mwader.png';
+import logo from '../../assets/images/Mwader.png';
 
 const Header = () => (
   <header className="header">
