+++ src/components/index.js
@@ -1,0 +1,2 @@
+export * from './common';
+export * from './ui';
