+++ src/components/ui/YesPage.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import Header from './Header';
-import Footer from './Footer';
+import { Header, Footer } from '../common';
 import './YesPage.css';
-import logo from '../Images/Mwader.png';
+import logo from '../../assets/images/Mwader.png';
 
 const products = [
   {
