+++ src/components/ui/index.js
@@ -1,0 +1,3 @@
+export { default as LostQ } from './LostQ';
+export { default as NoturePlace } from './NoturePlace';
+export { default as YesPage } from './YesPage';
