+++ src/pages/Yes/YesPage.css
@@ -1,0 +1,3 @@
+.yes-page {
+  min-height: 100vh;
+}
