+++ src/pages/Yes/YesPage.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { YesPage as YesPageComponent } from '../../components';
+import './YesPage.css';
+
+const YesPage = () => {
+  return (
+    <div className="yes-page">
+      <YesPageComponent />
+    </div>
+  );
+};
+
+export default YesPage;
