+++ src/pages/index.js
@@ -1,0 +1,7 @@
+export { default as HomePage } from './Home/HomePage';
+export { default as NoturePage } from './Noture/NoturePage';
+export { default as ProductsPage } from './Products/ProductsPage';
+export { default as ProductPage } from './Product/ProductPage';
+export { default as CartPage } from './Cart/CartPage';
+export { default as ContactPage } from './Contact/ContactPage';
+export { default as YesPage } from './Yes/YesPage';
